feat: publish day — security hardening, version cleanup, fresh testnet deploy
- Collapse v1-v8 deployment history to single TESTNET_PACKAGE_ID (0xbdbe...)
- Agent 1 security fixes: agent_mandate semantic 0/MAX, rate limiting,
  TOCTOU guard, fee validation in facilitator routes
- Agent 2 hardening: MCP format edge cases, test fixes
- Fresh testnet publish (10 modules, 264 Move tests passing)
- Update all docs with correct test counts (809 TS + 264 Move = 1,073)
- Remove all v7/v8 version references from public-facing docs

+++ contracts/sources/mandate.move
@@ -155,8 +155,13 @@ module sweefi::mandate {
         assert!(amount <= mandate.max_per_tx, EPerTxLimitExceeded);
 
         // Lifetime cap (unconditional — 0 means zero budget, NOT unlimited.
-        // This differs from AgentMandate where 0 = unlimited. See V8 audit F-14.)
-        assert!(mandate.total_spent + amount <= mandate.max_total, ETotalLimitExceeded);
+        // AgentMandate now matches this semantic after V8 audit F-19.)
+        // u128 intermediate: ensures EDailyLimitExceeded fires instead of generic
+        // ARITHMETIC_ERROR when total_spent + amount would overflow u64.
+        assert!(
+            (mandate.total_spent as u128) + (amount as u128) <= (mandate.max_total as u128),
+            ETotalLimitExceeded,
+        );
 
         // Debit
         mandate.total_spent = mandate.total_spent + amount;
